refactor: address AI review — dedup formatCompactDuration, add heatmap label
- Extract formatCompactDuration to dashboard-utils.ts (was defined identically
  in DashboardHome.tsx and InsightsPage.tsx)
- Add "Last 52 weeks" label to heatmap when a non-all range is selected,
  clarifying that the heatmap always shows full history
- Review issues #2 (tab visibility) was already correct — tab bar is wrapped
  in isEditor guard
