+++ packages/viewer/src/components/DashboardHome.tsx
@@ -2,6 +2,7 @@ import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import type { SessionSummary, SourceSession } from "../types";
 import { SessionDetailPopup } from "./Dashboard";
 import {
+  formatCompactDuration,
   isCacheFresh,
   navigateTo,
   normalizeTitleText,
@@ -40,18 +41,6 @@ interface InsightStats {
 }
 
 // ─── Helpers ─────────────────────────────────────────────────────────
-
-function formatCompactDuration(ms: number): string {
-  if (ms === 0) return "0m";
-  const mins = Math.floor(ms / 60000);
-  if (mins < 60) return `${mins}m`;
-  const hours = Math.floor(mins / 60);
-  const remMins = mins % 60;
-  if (hours < 24) return remMins > 0 ? `${hours}h ${remMins}m` : `${hours}h`;
-  const days = Math.floor(hours / 24);
-  const remHours = hours % 24;
-  return remHours > 0 ? `${days}d ${remHours}h` : `${days}d`;
-}
 
 // ─── Data Fetching ───────────────────────────────────────────────────
